pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "DriftcoordBot"
version = "0.1.0a1"
description = "A Discord bot with Chatterbox TTS voice cloning"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Dr1ftDev", email = "Dr1ftDev@github.com"}
]
license = {text = "MIT"}
keywords = ["discord", "bot", "discord.py", "tts", "chatterbox", "ai", "voice-cloning", "rocm", "cuda"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: No Input/Output (Daemon)",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

# Core dependencies (Base - CPU o Generico)
dependencies = [
    # Audio processing & ML
    "soundfile>=0.13.1",
    "librosa>=0.11.0",
    "scipy==1.11.4",
    "numpy>=1.25.2,<2.0.0",
    "pyworld==0.3.5",
    "praat-parselmouth==0.4.7",
    "pydub==0.25.1",
    "pedalboard==0.9.19",
    "demucs==4.0.1",
    "yt-dlp==2025.12.8",

    # Discord
    "discord.py>=2.6.4",
    "discord-ext-voice-recv==0.5.2a179",
    "PyNaCl==1.5.0",
    "python-dotenv>=1.2.1",

    # PyTorch ecosystem
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "torchvision>=0.21.0",

    # RVC Specific
    "omegaconf==2.0.6",
    "hydra-core==1.0.7",
    "fairseq==0.12.2",
    "rvc-python==0.1.0",
    "faiss-cpu==1.7.3",

    # AI/ML Extensions
    "transformers>=4.46.3,<4.50.0",
    "chatterbox-tts>=0.1.6",
    "peft==0.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "black>=25.12.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "mypy>=1.19.1",
    "ruff>=0.14.10",
]

# Supporto NVIDIA (CUDA 12.4+ come da requirements)
nvidia = [
    "torch==2.6.0",
    "torchaudio==2.6.0",
    "torchvision==0.21.0",
]

# Supporto AMD (ROCm 6.0/6.1)
amd = [
    "numpy<2.0.0",
    "cython>=3.0.0",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "torchvision>=0.21.0",
]

[project.urls]
Homepage = "https://github.com/DriftDeV/Python-DriftCoordBot"
Repository = "https://github.com/DriftDeV/Python-DriftCoordBot.git"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
